Added GLOBAL_DUPLICITY_CACHE_PARMS param in order to let the user configure where to keep the cache files. It can be used as follows: GLOBAL_DUPLICITY_CACHE_PARMS="--archive-dir=/home/backup/cache/duplicity" (issue #37)
Added SCP_PORT param in order to let the user configure the SFTP/SCP port when it is not the default 22. Also included the new SFTP related parameters, with the backup type as sftp (issue #38).

Removed the "--extra-clean" parameter, as this can led to problems when the user needs to restore a backup (issue #39)

Alfresco 4.2 is too old to be used as an example into the properties file. Changed the default to ALF_INSTALLATION_DIR=/opt/alfresco